Update unit tests for AB#8503
